#48 rework packages of gradle plugin
