+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/MultiplatformResourcesPlugin.kt
@@ -7,17 +7,17 @@ package dev.icerock.gradle
 import com.android.build.gradle.LibraryExtension
 import com.android.build.gradle.LibraryPlugin
 import com.android.build.gradle.api.AndroidSourceSet
-import dev.icerock.gradle.generator.AndroidMRGenerator
-import dev.icerock.gradle.generator.CommonMRGenerator
+import dev.icerock.gradle.generator.FontsGenerator
 import dev.icerock.gradle.generator.GenerateMultiplatformResourcesTask
-import dev.icerock.gradle.generator.IosMRGenerator
+import dev.icerock.gradle.generator.ImagesGenerator
 import dev.icerock.gradle.generator.MRGenerator
+import dev.icerock.gradle.generator.PluralsGenerator
 import dev.icerock.gradle.generator.ResourceGeneratorFeature
 import dev.icerock.gradle.generator.SourceInfo
-import dev.icerock.gradle.generator.fonts.FontsGeneratorFeature
-import dev.icerock.gradle.generator.image.ImagesGeneratorFeature
-import dev.icerock.gradle.generator.plurals.PluralsGeneratorFeature
-import dev.icerock.gradle.generator.strings.StringsGeneratorFeature
+import dev.icerock.gradle.generator.StringsGenerator
+import dev.icerock.gradle.generator.android.AndroidMRGenerator
+import dev.icerock.gradle.generator.common.CommonMRGenerator
+import dev.icerock.gradle.generator.ios.IosMRGenerator
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.tasks.SourceSet
@@ -82,10 +82,10 @@ class MultiplatformResourcesPlugin : Plugin<Project> {
         )
         val iosLocalizationRegion = mrExtension.iosBaseLocalizationRegion
         val features = listOf(
-            StringsGeneratorFeature(sourceInfo, iosLocalizationRegion),
-            PluralsGeneratorFeature(sourceInfo, iosLocalizationRegion),
-            ImagesGeneratorFeature(sourceInfo),
-            FontsGeneratorFeature(sourceInfo)
+            StringsGenerator.Feature(sourceInfo, iosLocalizationRegion),
+            PluralsGenerator.Feature(sourceInfo, iosLocalizationRegion),
+            ImagesGenerator.Feature(sourceInfo),
+            FontsGenerator.Feature(sourceInfo)
         )
         val targets: List<KotlinTarget> = multiplatformExtension.targets.toList()
