+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/BaseGenerator.kt
@@ -9,8 +9,6 @@ import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeSpec
-import dev.icerock.gradle.generator.strings.KeyType
-import dev.icerock.gradle.generator.strings.LanguageType
 import java.io.File
 
 abstract class BaseGenerator<T> : MRGenerator.Generator {
