+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/EqualStringKeysException.kt
@@ -2,7 +2,7 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.gradle.generator.strings
+package dev.icerock.gradle.generator
 
 class EqualStringKeysException(
     val keys: List<String>
