+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/ImagesGenerator.kt
@@ -1,15 +1,17 @@
 /*
- * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.gradle.generator.image
+package dev.icerock.gradle.generator
 
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.KModifier
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeSpec
-import dev.icerock.gradle.generator.MRGenerator
+import dev.icerock.gradle.generator.android.AndroidImagesGenerator
+import dev.icerock.gradle.generator.common.CommonImagesGenerator
+import dev.icerock.gradle.generator.ios.IosImagesGenerator
 import org.gradle.api.file.FileTree
 import java.io.File
 
@@ -60,4 +62,25 @@ abstract class ImagesGenerator(
     abstract fun getPropertyModifiers(): Array<KModifier>
 
     abstract fun getPropertyInitializer(key: String): CodeBlock?
+
+    class Feature(private val info: SourceInfo) : ResourceGeneratorFeature {
+        private val stringsFileTree = info.commonResources.matching {
+            include("MR/images/**/*.png", "MR/images/**/*.jpg")
+        }
+
+        override fun createCommonGenerator(): MRGenerator.Generator {
+            return CommonImagesGenerator(stringsFileTree)
+        }
+
+        override fun createIosGenerator(): MRGenerator.Generator {
+            return IosImagesGenerator(stringsFileTree)
+        }
+
+        override fun createAndroidGenerator(): MRGenerator.Generator {
+            return AndroidImagesGenerator(
+                stringsFileTree,
+                info.androidRClassPackage
+            )
+        }
+    }
 }
