+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/android/AndroidImagesGenerator.kt
@@ -1,12 +1,13 @@
 /*
- * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.gradle.generator.image
+package dev.icerock.gradle.generator.android
 
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.KModifier
+import dev.icerock.gradle.generator.ImagesGenerator
 import org.gradle.api.file.FileTree
 import java.io.File
