+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/common/CommonFontsGenerator.kt
@@ -2,10 +2,11 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.gradle.generator.fonts
+package dev.icerock.gradle.generator.common
 
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.KModifier
+import dev.icerock.gradle.generator.FontsGenerator
 import org.gradle.api.file.FileTree
 
 class CommonFontsGenerator(
