+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/common/CommonMRGenerator.kt
@@ -1,10 +1,11 @@
 /*
- * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.gradle.generator
+package dev.icerock.gradle.generator.common
 
 import com.squareup.kotlinpoet.KModifier
+import dev.icerock.gradle.generator.MRGenerator
 import org.gradle.api.Project
 import org.gradle.api.Task
 import java.io.File
