+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/common/CommonPluralsGenerator.kt
@@ -1,10 +1,11 @@
 /*
- * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package dev.icerock.gradle.generator.plurals
+package dev.icerock.gradle.generator.common
 
 import com.squareup.kotlinpoet.CodeBlock
+import dev.icerock.gradle.generator.PluralsGenerator
 import org.gradle.api.file.FileTree
 
 class CommonPluralsGenerator(
